feat: [website-file-breakdown-for-current-selection-005] - Extract work state selection package
