+++ ui/biome.jsonc
@@ -83,7 +83,7 @@
         "src/features/current-selection/work-selection/components/execution-details.test.tsx",
         "src/features/current-selection/components/provider-session-attempts.test.tsx",
         "src/features/current-selection/components/provider-session-detail-panel.test.tsx",
-        "src/features/current-selection/components/state-node-detail.test.tsx",
+        "src/features/current-selection/work-state-selection/components/state-node-detail.test.tsx",
         "src/features/current-selection/work-selection/components/work-item-card.test.tsx",
         "src/features/current-selection/workstation-selection/components/workstation-detail-card.editable-configuration.test.tsx",
         "src/features/current-selection/workstation-selection/components/workstation-detail-card.test.tsx",
