+++ ui/src/features/current-selection/components/detail-card-types.ts
@@ -2,3 +2,4 @@ export * from "../base/components/detail-card-types";
 export * from "../workstation-selection/lib/detail-card-types";
 export * from "../work-selection/lib/detail-card-types";
 export * from "../dispatch-selection/lib/detail-card-types";
+export * from "../work-state-selection/lib/detail-card-types";
