+++ ui/src/features/current-selection/work-state-selection/components/state-node-detail.test.tsx
@@ -1,9 +1,9 @@
 import { fireEvent, render, screen, within } from "@testing-library/react";
-import { semanticWorkflowDashboardSnapshot } from "../../../components/dashboard/test-fixtures";
-import { WIDGET_SUBTITLE_CLASS } from "../../../components/ui/widget-frame";
-import { formatTimeOfDay } from "../../../components/ui/formatters";
+import { semanticWorkflowDashboardSnapshot } from "../../../../components/dashboard/test-fixtures";
+import { WIDGET_SUBTITLE_CLASS } from "../../../../components/ui/widget-frame";
+import { formatTimeOfDay } from "../../../../components/ui/formatters";
 import { describe, expect, it, vi } from "vitest";
-import { CurrentSelectionLocaleProvider } from "./current-selection-locale";
+import { CurrentSelectionLocaleProvider } from "../../components/current-selection-locale";
 import { StateNodeDetailCard } from "./state-node-detail";
 
 function requireValue<T>(value: T | null | undefined, message: string): T {
