+++ ui/src/features/current-selection/work-state-selection/public/index.test.ts
@@ -1,0 +1,9 @@
+import { describe, expect, it } from "vitest";
+
+import * as workStateSelectionPublic from "./index";
+
+describe("work-state-selection/public", () => {
+  it("exports state node detail card and prop types", () => {
+    expect(workStateSelectionPublic.StateNodeDetailCard).toBeTypeOf("function");
+  });
+});
