+++ ui/src/features/current-selection/work-state-selection/public/index.ts
@@ -1,0 +1,7 @@
+export { StateNodeDetailCard } from "../components/state-node-detail";
+
+export type {
+  StateNodeDetailCardProps,
+  StatePositionWorkListItemProps,
+  StatePositionWorkListProps,
+} from "../lib/detail-card-types";
